requirements.txt
fastapi
pydantic
uvicorn
python-dotenv
pandas
streamlit
requests
openai
scikit-learn
faiss-cpu
transformers
langchain
sentence-transformers
pdfplumber
python-multipart